feat: #21 キーボード色分け
