+++ src/interfaces/AlphabetMatch.ts
@@ -1,0 +1,7 @@
+// アルファベットの判定を表す型を定義
+type LetterJudgement = 'Green' | 'Yellow' | 'Black' | 'NoUse';
+
+export interface AlphabetMatch {
+    [key: string]: LetterJudgement;
+}
+
